user dashboard moved, stubs for rfctr comps added
